requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
Flask-Migrate
Werkzeug
python-dotenv
PyMySQL
cryptography
requests
pandas
plotly
flask-cors


